requirements.txt
PyQt6==6.7.0
psycopg==3.2.1